chore: rearranging some files so they are easier to read
