+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/vegaprotocol/vega-snap.git"
   },
   "source": {
-    "shasum": "lITmlgEXyXWPrZ1pRIfYM1x4cxGWtBh4sayELtdlCCk=",
+    "shasum": "rqEo/ZVkVYcpnYYom6BUdTGJYpzj74CXv0p2AAZVFQk=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
