+++ packages/snap/src/index.ts
@@ -2,7 +2,7 @@ import { OnRpcRequestHandler, JsonRpcRequest } from '@metamask/snaps-sdk';
 import rpc from './node-rpc';
 import * as txs from './transaction';
 import { deriveKeyPair } from './keys';
-import { reviewTransaction } from './ui';
+import { reviewTransaction } from './transaction-ui';
 import {
   invalidParameters,
   JSONRPCError,
